services/event-consumer-logger/requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
